+++ moabb/datasets/base.py
@@ -361,6 +361,7 @@ class BaseDataset(metaclass=MetaclassDataset):
         paradigm,
         doi=None,
         unit_factor=1e6,
+        overlap=False,
     ):
         """Initialize function for the BaseDataset."""
         try:
@@ -390,6 +391,7 @@ class BaseDataset(metaclass=MetaclassDataset):
         self.paradigm = paradigm
         self.doi = doi
         self.unit_factor = unit_factor
+        self.overlap = overlap
 
     def _create_process_pipeline(self):
         return Pipeline(